Support vanilla Hyprland theme paths
- Add standalone Hyprland and pywal theme detection
- Share parsers between Hyprland and Omarchy providers
- Add coverage for vanilla config and Lua color formats
