+++ src/main/services/system-theme/providers/index.ts
@@ -1,5 +1,6 @@
 import type { DesktopThemeProvider } from '../system-theme.types'
 import { cosmicThemeProvider } from './cosmic'
+import { hyprlandThemeProvider } from './hyprland'
 import { kdeThemeProvider } from './kde'
 import { omarchyThemeProvider } from './omarchy'
 
@@ -7,4 +8,5 @@ export const desktopThemeProviders = [
   cosmicThemeProvider,
   kdeThemeProvider,
   omarchyThemeProvider,
+  hyprlandThemeProvider,
 ] satisfies DesktopThemeProvider[]
